fix(desktop+android): 全面代码审查修复 (安全/竞态/健壮性/i18n)
安全 (P0):
- auth.py: 登录日志只记 cookie 名不落值, 日志文件 0600 (原明文会话
  token 每秒写入公共临时目录)
- server.py: 静态文件服务 realpath 包含校验, 堵 Windows 盘符绝对路径
  任意读 (可读明文 token 数据库); 全请求 Host+Origin 本机校验
  (防 DNS rebinding 与跨站触发 logout/删除等破坏性操作)
- app.js: 检查更新弹窗远端 release tag 转义后再进 innerHTML
  (updater 版本尾缀不限制字符, 可注入标记)
- index.html: 加 CSP (script-src 'self'), 注入面降级为显示破坏

正确性/健壮性 (P1):
- db.py: monthly_cycle_start 对 end<=start 异常周期回退滚动 30 天
  (原死循环挂死请求线程); clear_account 级联清理 usage_charts;
  save_token 无活跃账号时兜底新增 (删光账号后 relogin 不再静默丢凭证);
  settings payload 读-改-写加锁防并发丢键; insert 冲突全列更新
  (含 account_id, 跨账号撞 usg_id 不再静默丢数据)
- main.py: 托盘配额面板修正缓存层级 (原恒为空); 登录成功 hide 不再
  产生孤儿窗口副作用
- server.py: /api/version 等三处补 return; 请求体限 1MiB + handler
  15s 读超时; 更新下载状态机加锁; 汇率改后台刷新不阻塞 dashboard;
  logout 丢弃在途 quota 刷新; quota 防重入原子化
- 前端: 同步轮询连续失败自动恢复按钮; 未登录停轮询风暴; 用户菜单
  序号守卫防幽灵弹出; 登录监视基线超时强制刷新; 设置页/首页加载
  失败有提示

维护性 (P2):
- updater.py: 按平台分流资产与完整性自检 (Windows 更新下载可用)
- 双 API 客户端: 响应超 4MiB 明确报错; GOAT 非法 createdAt 丢弃
- 前端: fmtMoney 负数/零统一; 分页封顶; 切账号复位筛选; i18n 补齐
  (timeout/GOAT 按钮/配额阶段/data-i18n-title); 弹窗 role/Esc 关闭,
  toast aria-live
- Android: allowBackup=false (明文 token 不进云备份); build.gradle
  支持 keystore.properties 正式签名 (无则回退 debug, CI 不变);
  每日统计缺天补 0 对齐桌面; 同步单飞加 Mutex; 登录完成清 WebView
  cookie; 响应有界读防 OOM; costRaw 改 Long (单条 >$21.47 溢出归零)

验证: 隔离库回归 9 项 + 登录状态机 6 场景全绿; Android gradle 编译
与单元测试通过。

+++ android/app/src/main/java/io/github/yphyphyph/gogauge/data/db/Entities.kt
@@ -40,7 +40,7 @@ data class UsageRecordEntity(
     @ColumnInfo(name = "cache_read_tokens") val cacheReadTokens: Int = 0,
     @ColumnInfo(name = "cache_write_5m_tokens") val cacheWrite5mTokens: Int = 0,
     @ColumnInfo(name = "cache_write_1h_tokens") val cacheWrite1hTokens: Int = 0,
-    @ColumnInfo(name = "cost_raw") val costRaw: Int = 0,
+    @ColumnInfo(name = "cost_raw") val costRaw: Long = 0,
     @ColumnInfo(name = "cost_usd") val costUsd: Double = 0.0,
     @ColumnInfo(name = "key_id") val keyId: String? = null,
     @ColumnInfo(name = "session_id") val sessionId: String? = null,
